Update README.md
Codimager is an innovative web application that empowers users to create visually appealing code images of your Flutter & Dart code that can be effortlessly shared on various social media sites. With its intuitive interface and a range of customization options, Codimager revolutionizes the way we present and share code on the internet.